People: Move Team to separate file
